world: fix data globals to match the original .data/.bss
Correct the size/type/value of several WORLD globals so their .data/.bss match
the original binary:
- ConfigBuffer: char * (4B) -> char[0x10000] (64KB buffer)
- CoinPacket: pointer + COINPACKETS[2] -> COINPACKET CoinPacket[2] (removes the
  extra COINPACKETS symbol)
- LevelTimer: TIMER* -> TIMER struct (0x10)
- DEFAULT_PLAYERHITPOINTS: u8=100 -> i32=8 (confirmed by comment in game.cpp)
- LEGOOBJ_DEFAULTLASTCOIN: u16=50 -> u32=-1
- Text_Language: 0 -> 1
- PlayerID: uninit -> {-1,-1}

Values verified against the original .data section.

+++ src/legoapi/world/level_shared.h
@@ -88,7 +88,7 @@ extern char AreaGlobals[0x34];
 extern i32 Door_UseCutCam;
 extern GameObject_s *Player[8];
 extern char PlayerProgress[0x80];
-extern u8 DEFAULT_PLAYERHITPOINTS;
+extern i32 DEFAULT_PLAYERHITPOINTS;
 extern i32 Area_PlayerModelCount;
 extern i32 Area_StoryModelCount;
 extern i16 Area_PlayerModelList[1];
